@@ -21,7 +21,11 @@ module Fae
     end
 
     version :thumb do
-      process :resize_to_fill => [150,100]
+      process :resize_to_fill => [150,100], if: :require_versions?
+    end
+
+    def require_versions?(new_file)
+      !['image/gif', 'image/svg+xml', 'text/plain'].include?(new_file.content_type)
     end
 
   end
