@@ -24,7 +24,7 @@ module Fae
 
       # Only allow a trusted parameter "white list" through.
       def option_params
-        params.require(:option).permit(:title, :time_zone, :colorway, :stage_url, :live_url, logo_attributes: [:id, :asset, :asset_cache, :attached_as, :alt], favicon_attributes: [:id, :asset, :asset_cache, :attached_as, :alt])
+        params.require(:option).permit(:title, :time_zone, :colorway, :stage_url, :live_url, logo_attributes: [:id, :asset, :asset_cache, :attached_as, :alt, :imageable_type, :required], favicon_attributes: [:id, :asset, :asset_cache, :attached_as, :alt, :imageable_type, :required])
       end
   end
 end
